Connected frontend to backend
Used anime.service.ts to call the api from the backend to web scrape. Then anime-list utilizes the service to call the api and display the information on the page. Cannot confirm it works yet, because urls.py and scraper.py have not been updated/created in the backend on this branch yet.