Add to Anime + Anime-Watchlist Tables
Anime can now be added to the database and to watchlists. This only happens in backend. Frontend does not show what anime are in which tables. Additionally, an anime can only be added to a watchlist if it is not in the database. If it is in the database, then it won't work.